% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/mark.R
\name{create_mark}
\alias{create_mark}
\alias{create_data_generic}
\alias{create_mark_generic}
\title{create_mark}
\usage{
create_data_generic(schema)

create_mark_generic(schema)

create_mark(mark, schema)
}
\arguments{
\item{schema}{}

\item{schema}{}
}
\description{
create_mark

create_mark
}
