% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{create_transform}
\alias{create_transform}
\title{create_transform}
\usage{
create_transform(trans, schema)
}
\arguments{
\item{schema}{imported json schema}

\item{transform}{name of transform, e.g 'BinTransform'}
}
\description{
create_transform
}
