% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema}
\alias{schema}
\alias{props}
\alias{enums}
\alias{types}
\title{schema}
\usage{
props(schema, type)

enums(schema, type)

types(schema, type)
}
\arguments{
\item{type}{}
}
\description{
schema utility functions
}
