% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{create_mark}
\alias{create_mark}
\alias{create_mark_generic}
\title{create_mark}
\usage{
create_mark_generic(schema)

create_mark(mark, schema)
}
\arguments{
\item{schema}{imported json schema}

\item{mark}{name of the mark}
}
\description{
create_mark
}
