% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema}
\alias{schema}
\alias{props}
\alias{enums}
\alias{types}
\title{schema}
\usage{
props(schema, type)

enums(schema, type)

types(schema, type)
}
\arguments{
\item{schema}{imported json schema}

\item{type}{schema reference}
}
\description{
schema utility functions
}
\examples{

schema_file <- Sys.glob(file.path(system.file("schema/vega-lite", package = "vegawidget"),"*.json"))
VL_SCHEMA <- jsonlite::read_json(schema_file)
encoding_options <- props(VL_SCHEMA, list("$ref" = "#/definitions/Encoding"))
}
